requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.4.0
attrs==24.2.0
av==12.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.4.0
decorator==4.4.2
diffusers==0.30.3
discord.py==2.4.0
distro==1.9.0
einops==0.8.0
fal_client==0.4.1
faster-whisper==1.0.3
filelock==3.16.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
iniconfig==2.0.0
Jinja2==3.1.3
jiter==0.5.0
MarkupSafe==2.1.5
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==2.1.1
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
onnxruntime==1.19.2
openai==1.45.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.10
protobuf==5.28.1
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-weather==2.0.3
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scipy==1.14.1
sniffio==1.3.1
sympy==1.13.2
tenacity==9.0.0
tokenizers==0.19.1
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==12.0
yarl==1.11.1
zipp==3.20.2
